AI_Sales_Bot-main/AI_Sales_Bot-main/requirements.txt
flask
python-dotenv
stripe
langchain-openai
fuzzywuzzy
python-Levenshtein
gunicorn
langchain
twilio
psycopg2-binary